requirements.txt
aniso8601==3.0.2
astroid==2.0.4
autopep8==1.3.5
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
Faker==0.9.0
flask==1.0.2
Flask-RESTful==0.3.6
Flask-SocketIO==3.0.1
gevent==1.3.6
gevent-websocket==0.10.1
greenlet==0.4.14
idna==2.7
isort==4.3.4
itsdangerous==0.24
jinja2==2.10
lazy-object-proxy==1.3.1
locustio==0.9.0.dev0
markupsafe==1.0
mccabe==0.6.1
msgpack==0.5.6
pycodestyle==2.4.0
pycparser==2.18
pylint==2.1.1
python-dateutil==2.7.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
pyzmq==17.1.2
requests==2.19.1
six==1.11.0
text-unidecode==1.2
typed-ast==1.1.0
urllib3==1.23
websocket==0.2.1
websocket-client==0.51.0
werkzeug==0.14.1
wrapt==1.10.11
